Bug 713542 - part 6: factor out Utils.join; remove SPECIAL_GUIDS nincompoopery; improve documentation.
